+++ app/Http/Controllers/TrafficController.php
@@ -5,31 +5,39 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Traffic;
 use App\Models\Intersection;
+use App\Models\ApiKey;
 
 class TrafficController extends Controller
 {
     public function index(Request $request){
         try {
             $status = 'success';
-            $message = '';
-            $data = Traffic::all(['id', 'name', 'address', 'vehiclesDensityInMinutes']);
+            if (ApiKey::where('key', $request->header('API_Key'))->count() > 0) {
+                $data = Traffic::all(['id', 'name', 'address', 'vehiclesDensityInMinutes']);
+            }else {
+                $status = 'failed';
+                $message = 'Access Denied';
+            }
         } catch (\Throwable $th) {
             //throw $th;
             $status = 'failed';
             $message = $th->getMessage();
-            $data = '';
         }
         return [
             "status" => $status,
-            "message" => $message,
-            "data" => $data
+            "message" => $message ?? '',
+            "data" => $data ?? ''
         ];
     }
     public function traffics(Request $request){
         try {
             $status = 'success';
-            $message = '';
-            $data = Traffic::take(5)->get(['id', 'name', 'address', 'vehiclesDensityInMinutes']);
+            if (ApiKey::where('key', $request->header('API_Key'))->count() > 0) {
+                $data = Traffic::take(5)->get(['id', 'name', 'address', 'vehiclesDensityInMinutes']);
+            }else {
+                $status = 'failed';
+                $message = 'Access Denied';
+            }
         } catch (\Throwable $th) {
             //throw $th;
             $status = 'failed';
@@ -38,17 +46,21 @@ class TrafficController extends Controller
         }
         return [
             "status" => $status,
-            "message" => $message,
-            "data" => $data
+            "message" => $message ?? '',
+            "data" => $data ?? ''
         ];
     }
     public function traffic_light_id($id){
         try {
             $status = 'success';
-            $message = '';
-            $data = Traffic::where('id', $id)->first(['id', 'name', 'address', 'vehiclesDensityInMinutes']);
-            $intersections = Intersection::where('traffic_id', $data->id)->get(['name', 'waitingTimeInSeconds', 'currentStatus']);
-            $data->intersections = $intersections;
+            if (ApiKey::where('key', $request->header('API_Key'))->count() > 0) {
+                $data = Traffic::where('id', $id)->first(['id', 'name', 'address', 'vehiclesDensityInMinutes']);
+                $intersections = Intersection::where('traffic_id', $data->id)->get(['name', 'waitingTimeInSeconds', 'currentStatus']);
+                $data->intersections = $intersections;
+            }else {
+                $status = 'failed';
+                $message = 'Access Denied';
+            }
         } catch (\Throwable $th) {
             //throw $th;
             $status = 'failed';
@@ -57,21 +69,25 @@ class TrafficController extends Controller
         }
         return [
             "status" => $status,
-            "message" => $message,
-            "data" => $data
+            "message" => $message ?? '',
+            "data" => $data ?? ''
         ];
     }
     public function traffic_store(Request $request){
         try {
             $status = 'success';
-            $message = '';
-            Traffic::create([
-                'name' => $request->name,
-                'address' => $request->address,
-                'latitude' => $request->latitude,
-                'longitude' => $request->longitude,
-                'vehiclesDensityInMinutes' => $request->vehiclesDensityInMinutes
-            ]);
+            if (ApiKey::where('key', $request->header('API_Key'))->count() > 0) {
+                Traffic::create([
+                    'name' => $request->name,
+                    'address' => $request->address,
+                    'latitude' => $request->latitude,
+                    'longitude' => $request->longitude,
+                    'vehiclesDensityInMinutes' => $request->vehiclesDensityInMinutes
+                ]);
+            }else {
+                $status = 'failed';
+                $message = 'Access Denied';
+            }
         } catch (\Throwable $th) {
             //throw $th;
             $status = 'failed';
@@ -79,20 +95,24 @@ class TrafficController extends Controller
         }
         return [
             "status" => $status,
-            "message" => $message
+            "message" => $message ?? ''
         ];
     }
     public function traffic_update(Request $request, $id){
         try {
             $status = 'success';
-            $message = '';
-            Traffic::where('id', $id)->update([
-                'name' => $request->name,
-                'address' => $request->address,
-                'latitude' => $request->latitude,
-                'longitude' => $request->longitude,
-                'vehiclesDensityInMinutes' => $request->vehiclesDensityInMinutes
-            ]);
+            if (ApiKey::where('key', $request->header('API_Key'))->count() > 0) {
+                Traffic::where('id', $id)->update([
+                    'name' => $request->name,
+                    'address' => $request->address,
+                    'latitude' => $request->latitude,
+                    'longitude' => $request->longitude,
+                    'vehiclesDensityInMinutes' => $request->vehiclesDensityInMinutes
+                ]);
+            }else {
+                $status = 'failed';
+                $message = 'Access Denied';
+            }
         } catch (\Throwable $th) {
             //throw $th;
             $status = 'failed';
@@ -100,15 +120,19 @@ class TrafficController extends Controller
         }
         return [
             "status" => $status,
-            "message" => $message
+            "message" => $message ?? ''
         ];
     }
     public function traffic_destroy($id){
         try {
             $status = 'success';
-            $message = '';
-            Traffic::where('id', $id)->delete();
-            Intersection::where('traffic_id', $id)->delete();
+            if (ApiKey::where('key', $request->header('API_Key'))->count() > 0) {
+                Traffic::where('id', $id)->delete();
+                Intersection::where('traffic_id', $id)->delete();
+            }else {
+                $status = 'failed';
+                $message = 'Access Denied';
+            }
         } catch (\Throwable $th) {
             //throw $th;
             $status = 'failed';
@@ -116,7 +140,7 @@ class TrafficController extends Controller
         }
         return [
             "status" => $status,
-            "message" => $message
+            "message" => $message ?? ''
         ];
     }
 }
